Fix some 64-bit MSVC warnings
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9089 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/aterm/source/aterm.cpp
@@ -2362,8 +2362,8 @@ static int AT_compareLists(ATermList t1, ATermList t2)
 
 int ATcompare(ATerm t1, ATerm t2)
 {
-  int type1;
-  int type2;
+  size_t type1;
+  size_t type2;
   int result = 0;
 
   if (ATisEqual(t1, t2))
